+++ app/models.py
@@ -118,15 +118,16 @@ class Dining(db.Model):
 
 class Student(db.Model):
     __tablename__ = 'students'
+    name = db.Column(db.String(64), nullable=False)
     uni = db.Column(db.String(8), unique=True, primary_key=True)
     email = db.Column(db.String(64), unique=True, nullable=False)
-    name = db.Column(db.String(64), nullable=False)
-    title = db.Column(db.String(128), nullable=True)
     department = db.Column(db.String(128), nullable=True)
+    title = db.Column(db.String(128), nullable=True)
     address = db.Column(db.String(512), nullable=False)
+    home_addr = db.Column(db.String(64), nullable=True)
+    campus_tel = db.Column(db.String(32), nullable=True)
     tel = db.Column(db.String(32), nullable=True)
     fax = db.Column(db.String(32), nullable=True)
-    home = db.Column(db.String(64), nullable=True)
 
     def __repr__(self):
         return f'<Student {self.uni}>'
